Contextqa client (#7)
* Add base gui structure

* Add `primevue` dependencies

* Add custom components

Added Chat and ChatCard components

* Add proxy

* Add `ContextManager` view

This view will set the context for a conversation session

* Add `MessageAdder` component

It handles the user questions and emits the corresponding
events so parent components can continue with the flow

* Update the `Chat` view

Now the messages are rendered dynamically

* Update dependencies

Added vue-router

* Add vue-router usage

* Update `ChatCard` component

Added datetime

* Update `ContextManager`

Added `remove` callback to the `FileUpload` component

* Update `ContextManager`

Added InputNumber fields

* Add `client` module and toast service

* Update client and server urls

* Update `ContextManager`

Added toast messages and the context request

* Add vuex dependency

* Add state management

* Add `askLLM` function

This is the main function to query the llm

* Update `Chat` component

Added support to query the llm server

* Update vuex store

* Update and add utilities to properly render the progress bar

* Add global font style

* Fix bug in state

the `lastMessageText` state was being rendered for all the
component instances.

* Update main input text

* Add autoscrolling and custom styles

* Fix bug in state

the identifier status changed just by selecting the file to upload.
Now it is only updated if the post request is made and succeeds

* Update `MessageAdder`

The keys shift+enter do not trigger the onEnter event

* Fix bug related to the cards sizes

* Update `ContextManager`

Now if the document has not been set as the context the chat
cannot be used

* Remove unused code

* Fix bug when the `Chat` component is re-created

* Remove the textarea outline when it's disabled

* Update `ChatCard`

Now it scans the LLM response to properly format code blocks

* Fix z-index bug

* Update the `client` module

Added validations to handle non-successful server responses

* Fix z-index bug

* Add logo and title

* Add `Home` view

* Update readme

* Update some strings

* Update `ContextManager`

Added dropdown to choose the vector store. Additionally, a new
state was added as the chosen vector store needs to be available
in the `Chat` component

* Add the `ConfirmationService` to use dialogs

* Add `ConfirmDialog` usage

For this to work, a setup script block is needed.

* Update the menubar

Added a custom "end" template to show the current context set

* Fix bugs

- `handleResponse` had not been awaited
-  invalid payload for the `lastMessageText` state

* Update project name

* Add a tooltip to the logo

* Add docker-related files for the client

* Update project structure

* Update query request

Now the processor is added from the state store

* Add API_BASE_URL for dev and prod

* Update api

- The source documents found are now logged when the
DEBUG flag is true
- Added cors middleware

* Update Dockerfiles

* Add docker compose files for dev and prod

* Update main script

- Removed previous commands
- Added new commands: `start` and `restart`

* Update api

Added `debug` setting

* Add commands starting messages

* Update readme

Added usage examples

* Update MenuBar chat options

* Add `chat` model

Added new endpoint to ask anything to the llm without
setting any document as context

* Add `Chat` and `DocumentQA` views

* Update `ChatBox`

Added `promise` method to request the corresponding
endpoint based on the context requirement

* Fix bugs related to messages state

* Update home message

* Update readme

* Update readme

* Update `Home`

Added contextqa title image

* Add `contextqa.env`

* Update contextqa image

* Update api

Renamed the `retriever` package to `contextqa`

* Update api settings

Added default values for optional settings

* Update chat cards

Now avatars have images instead of icons

* Fix bug

* Add focus to the textarea after a response

* Update menubar

Now the contextqa title is shown in the bar

* Update `ChatCard` sent date

* Update menubar

* Update readme

* Fix bug in the messages sent date

When the ChatBox was re-created, each message card displayed the
same sent date. This issue was resolved by changing the sent date
from being a computed property to a prop

* Renamed contextqa client root directory to **client**

* Add header in the `settings` section

* Fix `FileUpload` layout

* Update menubar

Now the set vector store is also shown

* Add docker restart strategy

* Update main script

Added `shutdown` command